Include Content-Length response header
Motivation:

The 'Content-Length' response header is expected by many clients.  While
it is not possible to specify the content length for on-the-fly
compression, it is possible when sending an individual file.

Modification:

Update DataSelection to optionally return the content length.

Update the 'getData' Response to include the `Content-Length` header
when the request targets a single file that is not placed in a zip
archive.  If the content is dynamically generated (e.g., a zip file)
then no `Content-Length` header is provided.

Update integration tests to verify correct behaviour.

Result:

IDS now includes the `Content-Length` response header when this is
possible

Signed-off-by: Paul Millar <paul.millar@desy.de>