Support testing larger files
Motivation:

Payara will behave differently depending on whether an HTTP response is
small enough to be cached; in particular, if the response is small
enough then, even if the content is dynamically generated, Payara will
not use chunked encoding.

Modification:

Add support for `FileRequirements` annotation, to identify a test that
needs special treatment.  Currently there is only one special treatment:
`BIG_DATAFILE`.  The presence of this flag indicates that the predefined
files needs to be big enough to force Payara to use chunked encoding.

Update code to verify zip and compressed content to allow verification
without knowing the compressed file's/zipped container's size.

Add tests that make use of this big-file testing, to demonstrate that
chunked encoding is taking place.

Result:

It is now possible to test IDS behaviour in scenarios that more closely
matches common data files.

Signed-off-by: Paul Millar <paul.millar@desy.de>